update :: test client setup 이동
