+++ test/integration/data/client/client_test.go
@@ -5,24 +5,14 @@ import (
 	"testing"
 
 	_ "github.com/mattn/go-sqlite3"
-	"github.com/pkg/errors"
 
 	"github.com/lucky-pocket/luckyPocket-back/internal/infra/data/client"
-	"github.com/lucky-pocket/luckyPocket-back/internal/infra/data/ent"
+	"github.com/lucky-pocket/luckyPocket-back/test/integration"
 	"github.com/stretchr/testify/assert"
 )
 
-func CreateTestClient() (c *ent.Client, closeFunc func(), err error) {
-	driver, dataSource := client.NewMemorySQLiteDialect()
-	c, closeFunc, err = client.NewClient(driver, dataSource)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "error createing test client")
-	}
-	return
-}
-
 func TestNewClient(t *testing.T) {
-	c, closeFunc, err := CreateTestClient()
+	c, closeFunc, err := integration.CreateTestClient()
 
 	if assert.NoError(t, err) {
 		defer closeFunc()
